ENH: Use HDFStore and TableIterator into contextmanagers
This adds `__enter__` and `__exit__` methods to both HDFStore and
TableIterator to make it easier to use them in with statements.

Also, this means you can use `HDFStore` more like a file object if you
want (i.e. wrap in a with statement so that you can be sure that enter
and exit happens). The only quirk is that HDFStore opens itself. To deal
with this case, get_store has HDFStore skip auto-opening, so that it can
be done in the `__enter__` method instead.

Also, `HDFStore` now takes a default warning level to use (e.g., if you
want to make sure that you don't accidentally overwrite a file, etc.)